Fix lerp_ dtype: use buffer dtype instead of hardcoded .float()
When model is bf16 (5090), buffers are bf16 too, so .float() creates
a mismatch in the opposite direction. Using .to(buffer.dtype) works
for both bf16 models and fp32 models with bf16 autocast.

+++ brain/dynamic_sparsity.py
@@ -132,7 +132,7 @@ class DynamicSparsityRouter(nn.Module):
         # Update EMA stats
         if self.training:
             with torch.no_grad():
-                self.avg_k_ema.lerp_(avg_k.detach().float(), 0.01)
+                self.avg_k_ema.lerp_(avg_k.detach().to(self.avg_k_ema.dtype), 0.01)
 
         stats = {
             "avg_active_experts": avg_k.item(),
@@ -242,7 +242,7 @@ class LateralInhibitionRouter(DynamicSparsityRouter):
         if self.training:
             with torch.no_grad():
                 self.steps_since_update += 1
-                self.avg_k_ema.lerp_(avg_k.detach().float(), 0.01)
+                self.avg_k_ema.lerp_(avg_k.detach().to(self.avg_k_ema.dtype), 0.01)
 
         stats = {
             "avg_active_experts": avg_k.item(),
